Finalize tests of the 2D drawable states. Fix a bug when change shapes from a track. The last shape will not change.
